refactor(web-ui): move helpers
