+++ packages/frontend/web-ui/src/auth/hooks/useLoginForm.ts
@@ -3,10 +3,10 @@ import { useEffect, useState } from 'react';
 import { useAppDispatch } from '../../app/store/hooks';
 import { createAuthByCredentials } from '../../app/store/thunk/createAuthByCredentials';
 import { isFullfilledPayloadAction } from '../../common/helpers/isFullfilledPayloadAction';
-import { validateEmail } from '../../common/helpers/validateEmail';
-import { validatePassword } from '../../common/helpers/validatePassword';
 import { OK, UNAUTHORIZED } from '../../common/http/helpers/httpCodes';
 import { Either } from '../../common/models/Either';
+import { validateEmail } from '../helpers/validateEmail';
+import { validatePassword } from '../helpers/validatePassword';
 import { FormValidationResult } from '../models/FormValidationResult';
 import {
   UseLoginFormParams,
